requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
openai>=1.3.0
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-core>=0.1.3
langchain-chroma>=0.0.5
chromadb>=0.4.18
unstructured>=0.10.30
pdf2image>=1.16.3
tqdm>=4.66.1
joblib>=1.3.2
PyPDF2>=3.0.0
pandas>=2.0.0

# Remove OCR/image/table processing dependencies
# The following dependencies are not needed for text-only processing:
# unstructured[all]
# unstructured-inference
# unstructured-pytesseract>=0.3.12
# pytesseract
# python-magic
# filetype
# pi-heif>=0.5.0

nltk
markdown

# Database dependencies
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
alembic>=1.13.0  # For database migrations